+++ jscomp/core/j.ml
@@ -76,7 +76,7 @@ and for_ident = ident
 and for_direction = Js_op.direction_flag
 and property_map = (property_name * expression) list
 and length_object = Js_op.length_object
-and delim = External_arg_spec.delim = | DNone | DStarJ | DJson
+and delim = External_arg_spec.delim = | DNone | DStarJ | DNoQuotes
 
 and expression_desc =
   | Length of expression * length_object
@@ -244,7 +244,7 @@ and case_clause = {
   comment : string option;
 }
 
-and string_clause = Lambda.as_value * case_clause
+and string_clause = Lambda.literal * case_clause
 and int_clause = int * case_clause
 
 and statement_desc =