+++ jscomp/core/lam_compile_const.ml
@@ -47,10 +47,10 @@ and translate (x : Lam_constant.t) : J.expression =
   | Const_js_false -> E.bool false
   | Const_js_null -> E.nil
   | Const_js_undefined -> E.undefined
-  | Const_int { i; comment = Pt_constructor {cstr_name={name; as_value=None}}} when name <> "[]" ->
+  | Const_int { i; comment = Pt_constructor {cstr_name={name; literal=None}}} when name <> "[]" ->
       E.str name
-  | Const_int { i; comment = Pt_constructor {cstr_name={as_value = Some as_value}}}  ->
-      E.as_value as_value
+  | Const_int { i; comment = Pt_constructor {cstr_name={literal = Some literal}}}  ->
+      E.literal literal
   | Const_int { i; comment } ->
       E.int i ?comment:(Lam_constant.string_of_pointer_info comment)
   | Const_char i -> Js_of_lam_string.const_char i