+++ jscomp/frontend/ast_attributes.ml
@@ -315,7 +315,7 @@ let iter_process_bs_string_or_int_as (attrs : Parsetree.attributes) =
                       | Some delim -> delim
                     in
                     st := Some (Str (s, delim));
-                    if delim = DJson then
+                    if delim = DNoQuotes then
                       (* check that it is a valid object literal *)
                       match
                         Classify_function.classify
@@ -335,7 +335,7 @@ let iter_process_bs_string_or_int_as (attrs : Parsetree.attributes) =
   !st
 
 let process_as_value (attrs : t) =
-  let st : Lambda.as_value option ref = ref None in
+  let st : Lambda.literal option ref = ref None in
   Ext_list.iter attrs (fun (({ txt; loc }, payload) as attr) ->
       match txt with
       | "bs.as" | "as" ->
@@ -344,33 +344,30 @@ let process_as_value (attrs : t) =
             | None -> ()
             | Some (s, _dec) ->
                 Bs_ast_invariant.mark_used_bs_attribute attr;
-                st := Some (AsString s));
+                st := Some (String s));
             (match Ast_payload.is_single_int payload with
             | None -> ()
             | Some i ->
                 Bs_ast_invariant.mark_used_bs_attribute attr;
-                st := Some (AsInt i));
+                st := Some (Int i));
             (match Ast_payload.is_single_float payload with
             | None -> ()
             | Some f ->
                 Bs_ast_invariant.mark_used_bs_attribute attr;
-                st := Some (AsFloat f));
+                st := Some (Float f));
             (match Ast_payload.is_single_bool payload with
             | None -> ()
             | Some b ->
                 Bs_ast_invariant.mark_used_bs_attribute attr;
-                st := Some (AsBool b));
+                st := Some (Bool b));
             (match Ast_payload.is_single_ident payload with
             | None -> ()
             | Some Lident "null" ->
                 Bs_ast_invariant.mark_used_bs_attribute attr;
-                st := Some AsNull
+                st := Some Null
             | Some Lident "undefined" ->
                 Bs_ast_invariant.mark_used_bs_attribute attr;
-                st := Some AsUndefined
-            | Some Lident "unboxed" ->
-                Bs_ast_invariant.mark_used_bs_attribute attr;
-                st := Some AsUnboxed
+                st := Some Undefined
             |  Some _ -> Bs_syntaxerr.err loc InvalidVariantAsAnnotation);
             if !st = None then Bs_syntaxerr.err loc InvalidVariantAsAnnotation
           )
@@ -394,6 +391,15 @@ let process_tag_name (attrs : t) =
           else Bs_syntaxerr.err loc Duplicated_bs_as
       | _ -> ());
   !st
+
+let process_untagged (attrs : t) =
+  let st = ref false in
+  Ext_list.iter attrs (fun (({ txt }, _)) ->
+      match txt with
+      | "unboxed" -> st := true
+      | _ -> ());
+  !st
+
 
 let locg = Location.none
 (* let bs : attr