+++ jscomp/frontend/ast_utf8_string_interp.ml
@@ -272,7 +272,7 @@ let transform_test s =
 module Delim = struct
   let parse_processed = function
     | None -> Some External_arg_spec.DNone
-    | Some "json" -> Some DJson
+    | Some "json" -> Some DNoQuotes
     | Some "*j" -> Some DStarJ
     | _ -> None