+++ jscomp/ml/lambda.ml
@@ -38,8 +38,6 @@ type record_repr =
   | Record_regular
   | Record_optional
 
-type as_value = AsString of string | AsInt of int | AsFloat of string | AsBool of bool | AsNull | AsUndefined | AsUnboxed
-type cstr_name = {name: string; as_value: as_value option}
 
 type tag_info = 
   | Blk_constructor of {name : string ; num_nonconst : int ; tag : int; attrs : Parsetree.attributes }
@@ -273,7 +271,14 @@ type function_attribute = {
   return_unit : bool;
   async : bool;
 }
-type block = {cstr_name: cstr_name; tag_name: string option}
+
+type block_type =
+  | IntType | StringType | FloatType | Array | Object | Unknown
+type literal =
+  | String of string | Int of int | Float of string | Bool of bool | Null | Undefined
+  | Block of block_type
+type cstr_name = {name: string; literal: literal option}
+type block = {cstr_name: cstr_name; tag_name: string option; block_type : block_type option}
 type switch_names = {consts: cstr_name array; blocks: block array}
 
 type lambda =