+++ jscomp/ml/matching.ml
@@ -1329,12 +1329,10 @@ let make_constr_matching p def ctx = function
     [] -> fatal_error "Matching.make_constr_matching"
   | ((arg, _mut) :: argl) ->
       let cstr = pat_as_constr p in
+      let untagged =
+        Ext_list.exists cstr.cstr_attributes (function ({txt}, _) -> txt = "unboxed") in
       let newargs =
-        if cstr.cstr_inlined <> None ||
-          Ext_list.exists cstr.cstr_attributes (function
-            | ({txt="as"}, PStr [{pstr_desc = Pstr_eval
-                ({pexp_desc = Pexp_ident {txt= Lident "unboxed"}}, _)}]) -> true
-            | _ -> false) then
+        if cstr.cstr_inlined <> None || (untagged && cstr.cstr_args <> []) then
           (arg, Alias) :: argl
         else match cstr.cstr_tag with
         | Cstr_block _ when