+++ jscomp/test/variantsMatching.gen.tsx
@@ -7,3 +7,15 @@ export type t = "thisIsA" | 42 | null | "D" | 3.14;
 
 // tslint:disable-next-line:interface-over-type-literal
 export type tNU = null | undefined;
+
+// tslint:disable-next-line:interface-over-type-literal
+export type MyUndefined_t<a> = undefined | a;
+
+// tslint:disable-next-line:interface-over-type-literal
+export type MyNull_t<a> = null | a;
+
+// tslint:disable-next-line:interface-over-type-literal
+export type MyNullable_t<a> = null | undefined | a;
+
+// tslint:disable-next-line:interface-over-type-literal
+export type MyNullableExtended_t<a> = null | undefined | "WhyNotAnotherOne" | a;